BC-12022 - move some interfaces / enums back to file record DO
