+++ src/modules/files-storage/domain/contract/file-record-status.interface.ts
@@ -1,5 +1,5 @@
 import { type ScanStatus } from '../vo';
-import { type PreviewStatus } from './preview-status.enum';
+import { type PreviewStatus } from '../file-record.do';
 
 export interface FileRecordStatus {
 	scanStatus: ScanStatus;
