+++ src/modules/files-storage/domain/contract/file-record-with-status.interface.ts
@@ -1,5 +1,5 @@
 import { type FileRecord } from '../file-record.do';
-import { type FileRecordStatus } from '../interface';
+import { type FileRecordStatus } from './file-record-status.interface';
 
 export interface FileRecordWithStatus {
 	fileRecord: FileRecord;
