+++ src/modules/files-storage/domain/contract/index.ts
@@ -1,3 +1,5 @@
+export { FileRecordStatus } from './file-record-status.interface';
 export { FileRecordWithStatus } from './file-record-with-status.interface';
 export { FILE_RECORD_REPO, FileRecordRepo } from './file-record.repo.interface';
 export { PreviewFileParams } from './preview-file-params.interface';
+export { PreviewInfo } from './preview-info.interface';
