+++ src/modules/files-storage/domain/contract/preview-info.interface.ts
@@ -1,0 +1,8 @@
+import { type PreviewOutputMimeTypes } from '../file-record.do';
+import { type PreviewWidth } from '../interface';
+
+export interface PreviewInfo {
+	outputFormat?: PreviewOutputMimeTypes;
+	width?: PreviewWidth;
+	forceUpdate?: boolean;
+}
