+++ src/modules/files-storage/domain/index.ts
@@ -1,18 +1,40 @@
-export * from './contract';
+export {
+	FILE_RECORD_REPO,
+	FileRecordRepo,
+	FileRecordStatus,
+	FileRecordWithStatus,
+	PreviewFileParams,
+	PreviewInfo,
+} from './contract';
 export { FileDto } from './dto';
 export { ErrorType } from './error';
 export { FileDtoFactory, FileRecordFactory, ParentStatisticFactory } from './factory';
-export * from './file-record.do';
-export * from './interface';
+export {
+	FileRecord,
+	FileRecordProps,
+	PreviewOutputMimeTypes,
+	PreviewStatus,
+	TEMP_FILE_EXPIRY_SECONDS,
+} from './file-record.do';
+export {
+	CopyFileResult,
+	DocumentType,
+	FileRecordParentType,
+	GetFileResponse,
+	ParentInfo,
+	ParentReference,
+	PreviewWidth,
+	StorageLocation,
+	StorageLocationParams,
+} from './interface';
 export { FileStorageActionsLoggable } from './loggable';
 export { FilesStorageMapper } from './mapper';
 export { FilesStorageService, PreviewService } from './service';
-export * from './storage-paths.const';
+export { FolderExpirationDays, StorageFolders, StorageType } from './storage-paths.const';
 export {
 	FileRecordSecurityCheck,
 	FileRecordSecurityCheckProps,
 	ParentStatistic,
 	ParentStatisticProps,
-	ScanStatus
+	ScanStatus,
 } from './vo';
-
