--- src/modules/files-storage/domain/interface/preview-info.interface.ts
@@ -1,8 +1,0 @@
-import { type PreviewOutputMimeTypes } from './preview-output-mime-types.enum';
-import { type PreviewWidth } from './preview-width.enum';
-
-export interface PreviewInfo {
-	outputFormat?: PreviewOutputMimeTypes;
-	width?: PreviewWidth;
-	forceUpdate?: boolean;
-}
