--- src/modules/files-storage/domain/interface/preview-status.enum.ts
@@ -1,8 +1,0 @@
-export enum PreviewStatus {
-	PREVIEW_POSSIBLE = 'preview_possible',
-	AWAITING_SCAN_STATUS = 'awaiting_scan_status',
-	PREVIEW_NOT_POSSIBLE_SCAN_STATUS_ERROR = 'preview_not_possible_scan_status_error',
-	PREVIEW_NOT_POSSIBLE_SCAN_STATUS_WONT_CHECK = 'preview_not_possible_scan_status_wont_check',
-	PREVIEW_NOT_POSSIBLE_SCAN_STATUS_BLOCKED = 'preview_not_possible_scan_status_blocked',
-	PREVIEW_NOT_POSSIBLE_WRONG_MIME_TYPE = 'preview_not_possible_wrong_mime_type',
-}
