+++ src/modules/files-storage/domain/service/files-storage-delete.service.spec.ts
@@ -14,8 +14,8 @@ import { fileRecordTestFactory } from '../../testing';
 import { FILE_RECORD_REPO, type FileRecordRepo } from '../contract';
 import { ErrorType } from '../error';
 import { FilePathFactory } from '../factory';
-import { FileRecord } from '../file-record.do';
-import { FileRecordProps, StorageLocation } from '../interface';
+import { FileRecord, type FileRecordProps } from '../file-record.do';
+import { StorageLocation } from '../interface';
 import { StorageLocationDeleteLoggableException } from '../loggable';
 import { type FileRecordSecurityCheckProps } from '../vo';
 import { FilesStorageService } from './files-storage.service';
