+++ src/modules/files-storage/domain/service/files-storage-restore.service.spec.ts
@@ -12,8 +12,7 @@ import {
 import { fileRecordTestFactory, ParentInfoTestFactory } from '../../testing';
 import { FILE_RECORD_REPO, type FileRecordRepo } from '../contract';
 import { FilePathFactory, FileRecordFactory } from '../factory';
-import { FileRecord } from '../file-record.do';
-import { type FileRecordProps } from '../interface';
+import { FileRecord, type FileRecordProps } from '../file-record.do';
 import { FileRecordSecurityCheck, type FileRecordSecurityCheckProps } from '../vo';
 import { FilesStorageService } from './files-storage.service';
